+++ lib/cms/authentication/test_password_strategy.rb
@@ -17,3 +17,12 @@ module Cms
 end
 
 Warden::Strategies.add(:test_password, Cms::Authentication::TestPasswordStrategy)
+
+# NOTE:  To enable a custom password strategy for BrowserCMS you must also add it to to the devise configuration.
+# 
+# For example enable the test_password strategy above by adding the following to config/initializers/devise.rb
+
+# # Add test_password strategy BEFORE other CMS authentication strategies
+# config.warden do |manager|
+#   manager.default_strategies(:scope => :cms_user).unshift :test_password
+# end