+++ app/helpers/cms/content_types_helper.rb
@@ -1,0 +1,4 @@
+module Cms
+  module ContentTypesHelper
+  end
+end