+++ app/portlets/helpers/cms/list_portlet_helper.rb
@@ -1,0 +1,5 @@
+##
+# All methods from this helper will be available in the render.html.erb for ListPortlet
+module ListPortletHelper
+  
+end