+++ lib/generators/cms/portlet/templates/portlet.rb
@@ -1,5 +1,7 @@
 class <%= class_name %>Portlet < Cms::Portlet
 
+  description "TODO: Provide a suitable description for this portlet."
+
   # Mark this as 'true' to allow the portlet's template to be editable via the CMS admin UI.
   enable_template_editor false