+++ test/helpers/cms/content_types_helper_test.rb
@@ -1,0 +1,9 @@
+require "test_helper"
+
+describe ContentTypesHelper do
+
+  it "must be a real test" do
+    flunk "Need real tests"
+  end
+
+end